server: get default port optionally form environment var
If the default port 8080 is taken, you have to either change the app.run()
or  use --port <PORTNUM> on every invocation of the program. When trying
out examples I don't want to do the first and forget the second.

This patch allows you set the environment variables LONA_DEFAULT_HOST and
LONA_DEFAULT_PORT in your shell. These override the built-in defaults
('localhost' resp. 8080), and are in turn overridden by arguments
to ``app.run()`` and startup commandline parameters.